Criada função criptografia e função pra ler texto

+++ Headers/funcoesMatriz.h
@@ -3,9 +3,9 @@
 
 #define MAX 2
 
-void gerarMat(int mat[MAX][MAX]);
-int calcDet(int mat[MAX][MAX]);
-void calcMatAdj(int mat[MAX][MAX], float matAdj[MAX][MAX]);
+void gerarMat(float mat[MAX][MAX]);
+int calcDet(float mat[MAX][MAX]);
+void calcMatAdj(float mat[MAX][MAX], float matAdj[MAX][MAX]);
 void calcMatInv(float matAdj[MAX][MAX], float matInv[MAX][MAX], int det);
 
 #endif // FUNCOESMATRIZ_H
